keggConv {KEGGREST}	R Documentation

Convert KEGG identifiers to/from outside identifiers

Description

Convert KEGG identifiers to/from outside identifiers.

Usage

keggConv(target, source, querySize = 100)

Arguments

target
A KEGG organism code (), T number, or one of the external databases ncbi-gi, ncbi-geneid, ncbi-proteinid, uniprot, or (for chemical substance identifiers) drug, compound, or glycan, pubchem, or chebi.

source
Same as target, but may also be a list of KEGG identifers representing internal or external names.

querySize
Empirically, KEGG limits queries to 100 source identifiers per query. This argument enables larger queries by dividing source into sub-queries of no more than querySize identifiers.

Value

A named character vector.

Author(s)

Dan Tenenbaum

References

https://www.kegg.jp/kegg/docs/keggapi.html

Examples

## conversion from NCBI GeneID to KEGG ID for E. coli genes
head(keggConv("eco", "ncbi-geneid"))
head(keggConv("ncbi-geneid", "eco")) ## opposite direction

## conversion from KEGG ID to NCBI GI
head(keggConv("ncbi-proteinid", c("hsa:10458", "ece:Z5100")))

## conversion from NCBI GI to KEGG ID when the organism code is not known:
head(keggConv("genes", "ncbi-geneid:3113320"))

[Package KEGGREST version 1.34.0 Index]