keggGet {KEGGREST}	R Documentation

Retrieves given database entries

Description

Retrieves given database entries.

Usage

keggGet(dbentries, option = c("aaseq", "ntseq", "mol", "kcf", 
    "image", "kgml"))

Arguments

dbentries
One or more (up to a maximum of 10) KEGG identifiers.

option
Optional. Option governing the format of the output. aaseq is an amino acid sequence, ntseq is a nucleotide sequence. image returns an object which can be written to a PNG file, kgml returns a KGML document.

Details

Retrieves all entries from the KEGG database for a set of KEGG identifers.

keggGet() can only return 10 result sets at once (this limitation is on the server side). If you supply more than 10 inputs to keggGet(), KEGGREST will warn that only the first 10 results will be returned.

Value

A list wrapping a KEGG flat file. If option is aaseq, an AAStringSet object. If option is ntseq, a DNAStringSet object. If option is image, an object which can be written to a PNG file. If option is kgml, a KGML document.

Author(s)

Dan Tenenbaum

References

https://www.kegg.jp/kegg/docs/keggapi.html

Examples

res <- keggGet(c("cpd:C01290", "gl:G00092")) ## retrieves a compound entry
                                    ## and a glycan entry
str(res)
res <- keggGet(c("C01290", "G00092")) ## same as above, without prefixes
str(res)
res <- keggGet(c("hsa:10458", "ece:Z5100")) ## retrieves a human gene entry
                                    ## and an E.coli O157 gene entry
str(res)
res <- keggGet(c("hsa:10458", "ece:Z5100"), "aaseq") ## retrieves amino
                                    ## acid sequences of a human gene and an 
                                    ## E.coli O157 gene
png <- keggGet("hsa05130", "image") ## retrieves the image file of a
                                    ## pathway map
t <- tempfile()
library(png)
writePNG(png, t)
res <- keggGet("hsa05130", "kgml")
str(res)

[Package KEGGREST version 1.34.0 Index]