keggList {KEGGREST}	R Documentation

Returns a list of entry identifiers and associated definition for a given database or a given set of database entries.

Description

Returns a list of entry identifiers and associated definition for a given database or a given set of database entries.

Usage

keggList(database, organism)

Arguments

database
Either a KEGG database (list available via listDatabases()), a KEGG organism code (list available via keggList() with the organism argument, a T number (list available via keggList() with the genome argument), or a character vector of KEGG identifiers.

organism
Optional. A KEGG organism identifier (list available via keggList() with the organism argument).

Value

A named character vector containing entry identifiers and associated definition.

Author(s)

Dan Tenenbaum

References

https://www.kegg.jp/kegg/docs/keggapi.html

Examples

res <- keggList("pathway") ## returns the list of reference pathways
length(res)
head(res) 
res <- keggList("pathway", "hsa") ## returns the list of human pathways
length(res)
head(res)
res <- keggList("organism") ## returns the list of KEGG organisms with
                     ## taxonomic classification
nrow(res)
head(res)
res <- keggList("hsa")  ## returns the entire list of human genes
length(res)
head(res)
## keggList("T01001") ## same as above
keggList(c("hsa:10458", "ece:Z5100")) ## returns the list of a human gene
                                      ## and an E.coli O157 gene
keggList(c("cpd:C01290","gl:G00092")) ## returns the list of a compound entry
                                      ## and a glycan entry
keggList(c("C01290+G00092")) ## same as above (prefixes are not necessary)

[Package KEGGREST version 1.34.0 Index]